server/requirements.txt
fastapi==0.135.3
uvicorn==0.42.0
pydantic==2.12.5
openenv-core>=0.2.0
